+++ app/controllers/teams_controller.rb
@@ -7,7 +7,10 @@ class TeamsController < ApplicationController
     respond_to do |format|
       format.html
       format.json do
-        render json: @presenter.to_json(graphs: [{ type: :line_graph, data: :team_data, units: '%' }])
+        dataset_labels = t("graphs.#{params[:graph_kind]}.dataset_labels", default: nil)
+        units = t("graphs.#{params[:graph_kind]}.units", default: '')
+
+        render json: @presenter.to_json(graphs: [{ type: :line_graph, data: :team_data, units: units, dataset_labels: dataset_labels }])
       end
     end
   end
@@ -35,7 +38,7 @@ class TeamsController < ApplicationController
 
   def team_params
     params.permit(:id, :format, :page, :user_ids, :plan_id, :actual_id,
-                  :filter_start_month, :filter_end_month, :filter_start_year,
-                  :filter_end_year, :filter_tag_ids)
+                  :filter_start_month, :filter_end_month, :filter_start_year, :filter_end_year, :filter_tag_ids,
+                  :graph_kind)
   end
 end