+++ app/models/time_range.rb
@@ -30,12 +30,6 @@ class TimeRange < ApplicationRecord
     (value || 0) * 60.0
   end
 
-  def validate_end_time_after_start_time
-    return unless start_time && end_time && end_time <= start_time
-
-    errors.add :end_time, 'must occur after start time'
-  end
-
   def segment_value(segment_start:, segment_end:)
     value * Intersection.call(
       a_start: start_time,
@@ -43,5 +37,13 @@ class TimeRange < ApplicationRecord
       b_start: segment_start,
       b_end: segment_end
     )
+  end
+
+  private
+
+  def validate_end_time_after_start_time
+    return unless start_time && end_time && end_time <= start_time
+
+    errors.add :end_time, 'must occur after start time'
   end
 end