+++ spec/api/v1/time_ranges/create_spec.rb
@@ -11,7 +11,7 @@ describe Api::V1::TimeRangeResource, type: :request, swagger_doc: 'v1/swagger.js
       user_id: user.id,
       start_time: 1.hour.since.iso8601,
       end_time: 20.hours.since.iso8601,
-      seconds_worked: (24 * 3600).to_s,
+      minutes_worked: 50,
       time_range_type_id: time_range_type.id
     }
   end
@@ -38,10 +38,12 @@ describe Api::V1::TimeRangeResource, type: :request, swagger_doc: 'v1/swagger.js
       let(:Authorization) { 'Bearer dummy_json_web_token' }
 
       response '201', 'Time Range created' do
-        schema '$ref' => '#/definitions/time_range_post_params'
+        schema '$ref' => '#/definitions/time_range_response'
 
         run_test! do
-          parsed_json_data_matches_db_record(created_time_range)
+          data = parsed_json_data
+          data['attributes']['value'] = data['attributes'].delete('minutes_worked').to_d.to_s
+          parsed_json_data_matches_db_record(created_time_range, data)
         end
       end
 
@@ -60,6 +62,41 @@ describe Api::V1::TimeRangeResource, type: :request, swagger_doc: 'v1/swagger.js
         response '422', 'Invalid request' do
           schema '$ref' => '#/definitions/error_422'
           run_test!
+        end
+      end
+
+      context 'tag relationships passed' do
+        let!(:tag1) { create :tag }
+        let!(:tag2) { create :tag }
+        let!(:tag3) { create :tag }
+
+        let(:relationships) do
+          {
+            tags: {
+              data: [
+                { type: 'tags', id: tag2.id },
+                { type: 'tags', id: tag3.id }
+              ]
+            }
+          }
+        end
+
+        let(:time_range) do
+          {
+            data: {
+              type: 'time_ranges',
+              attributes: attributes,
+              relationships: relationships
+            }
+          }
+        end
+
+        response '201', 'Time Range created' do
+          schema '$ref' => '#/definitions/time_range_response_with_relationships'
+
+          run_test! do
+            expect(created_time_range.tags.collect(&:id)).to match_array [tag2.id, tag3.id]
+          end
         end
       end
     end