+++ spec/api/v1/time_ranges/index_spec.rb
@@ -4,8 +4,8 @@ describe Api::V1::TimeRangeResource, type: :request, swagger_doc: 'v1/swagger.js
   let(:user1) { create :user }
   let(:user2) { create :user }
 
-  let(:time_range_type_name1) { Faker::Lorem.word }
-  let(:time_range_type_name2) { Faker::Lorem.word }
+  let(:time_range_type_name1) { Faker::Lorem.unique.word }
+  let(:time_range_type_name2) { Faker::Lorem.unique.word }
 
   let(:time_range_type1) { create :time_range_type, name: time_range_type_name1 }
   let(:time_range_type2) { create :time_range_type, name: time_range_type_name2 }
@@ -22,27 +22,54 @@ describe Api::V1::TimeRangeResource, type: :request, swagger_doc: 'v1/swagger.js
       parameter name: 'page[number]', in: :query, type: :integer, required: false
       parameter name: 'filter[user_id]', in: :query, type: :integer, required: false
       parameter name: 'filter[time_range_type_id]', in: :query, type: :integer, required: false
+      parameter name: 'include', in: :query, type: :string, required: false
 
-      let!(:'page[size]') { 10 }
-      let!(:'page[number]') { 1 }
+      let(:include) { '' }
+
+      let(:'page[size]') { 10 }
+      let(:'page[number]') { 1 }
 
       let(:Authorization) { 'Bearer dummy_json_web_token' }
 
       response '200', 'Time Ranges index' do
         schema '$ref' => '#/definitions/time_ranges_response'
 
-        describe 'attributes match database values' do
-          run_test! do
-            expect(parsed_json_data.count).to eq(2)
+        run_test! do
+          expect(parsed_json_data.count).to eq(2)
 
-            database_record = TimeRange.find(parsed_json_data.first['id'])
-            parsed_json_data_matches_db_record(database_record, parsed_json_data.first)
+          database_record = TimeRange.find(parsed_json_data.first['id'])
+
+          data = parsed_json_data.first
+          data['attributes']['value'] = data['attributes'].delete('minutes_worked').to_d.to_s
+
+          parsed_json_data_matches_db_record(database_record, data)
+        end
+
+        context 'include tags' do
+          let(:include) { 'tags' }
+
+          let!(:parent_tag_type) { create :tag_type }
+          let!(:tag_type) { create :tag_type, parent: parent_tag_type }
+          let!(:parent_tag) { create :tag, tag_type: parent_tag_type }
+
+          let!(:tag1) { create :tag, parent: parent_tag, tag_type: tag_type }
+          let!(:tag2) { create :tag, parent: parent_tag, tag_type: tag_type }
+          let!(:tag3) { create :tag, parent: parent_tag, tag_type: tag_type }
+
+          let!(:tag_association1) { create :tag_association, :skip_validate, tag: tag2, taggable: time_range1, tag_type: parent_tag_type }
+          let!(:tag_association2) { create :tag_association, :skip_validate, tag: tag3, taggable: time_range1, tag_type: parent_tag_type }
+
+          run_test! do
+            expect(parsed_json_data.first['relationships']['tags']['data'].collect { |e| e['id'].to_i }).to match_array [tag2.id, tag3.id]
+            expect(parsed_json['included'].collect { |e| e['attributes']['name'] }).to match_array [tag2.name, tag3.name]
+
+            expect(parsed_json_data.second['relationships']['tags']['data']).to be_empty
           end
         end
 
         context 'with 1 result per page' do
-          let!(:'page[size]') { 1 }
-          let!(:'page[number]') { 2 }
+          let(:'page[size]') { 1 }
+          let(:'page[number]') { 2 }
 
           run_test! do
             expect(parsed_json_data.first['id']).to eq(time_range2.id.to_s)