+++ swagger/v1/time_ranges.rb
@@ -6,11 +6,11 @@ module Swagger
           time_range_attributes: {
             type: 'object',
             properties: {
-              start_time: { type: 'string', example: Date.parse('2021-01-22').beginning_of_day.iso8601, 'x-nullable': false },
-              end_time: { type: 'string', example: Date.parse('2021-01-24').end_of_day.iso8601, 'x-nullable': false },
+              start_time: { type: 'string', example: DateTime.parse('2021-01-22 09:00:00').iso8601, 'x-nullable': false },
+              end_time: { type: 'string', example: DateTime.parse('2021-01-22 09:45:00').iso8601, 'x-nullable': false },
               user_id: { type: 'integer', example: 123, 'x-nullable': false },
-              seconds_worked: { type: 'string', example: (3600 * 16).to_s, 'x-nullable': false },
-              time_range_type_id: { type: 'integer', example: 567, 'x-nullable': false }
+              minutes_worked: { type: 'integer', example: 45, 'x-nullable': false },
+              time_range_type_id: { type: 'integer', example: 1, 'x-nullable': false }
             }
           },
           time_ranges_response: {
@@ -67,6 +67,44 @@ module Swagger
               }
             }
           },
+          time_range_response_with_relationships: {
+            type: 'object',
+            properties: {
+              data: {
+                type: 'object',
+                properties: {
+                  id: { type: 'string', example: '1' },
+                  type: { type: 'string', example: 'time_ranges' },
+                  link: {
+                    type: 'object',
+                    properties: {
+                      self: { type: 'string', example: 'https://job-plan-stats.herokuapp.com/api/v1/time_ranges/1' }
+                    }
+                  },
+                  attributes: { '$ref' => '#/definitions/time_range_attributes' },
+                  relationships: {
+                    type: 'object',
+                    properties: {
+                      tags: {
+                        properties: {
+                          data: {
+                            type: 'array',
+                            items: {
+                              type: 'object',
+                              properties: {
+                                id: { type: 'string', example: '105' },
+                                type: { type: 'string', example: 'tags' }
+                              }
+                            }
+                          }
+                        }
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          },
           time_range_post_params: {
             type: 'object',
             properties: {
@@ -74,7 +112,26 @@ module Swagger
                 type: 'object',
                 properties: {
                   type: { type: 'string', example: 'time_ranges' },
-                  attributes: { '$ref' => '#/definitions/time_range_attributes' }
+                  attributes: { '$ref' => '#/definitions/time_range_attributes' },
+                  relationships: {
+                    type: 'object',
+                    properties: {
+                      tags: {
+                        properties: {
+                          data: {
+                            type: 'array',
+                            items: {
+                              type: 'object',
+                              properties: {
+                                id: { type: 'string', example: '105' },
+                                type: { type: 'string', example: 'tags' }
+                              }
+                            }
+                          }
+                        }
+                      }
+                    }
+                  }
                 }
               }
             }