+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.8] - 2021-02-05
+
+### Added
+
+- Planned vs Actual on Team Dashboard line graph
+
+## [0.9.7] - 2021-02-04
+
+### Added
+
+- TagType and Tag API endpoints
+
 ## [0.9.6] - 2021-01-26
 
 ### Added